setState before it is unmounted
When it close the lightbox , there will be a warning, "Warning: setState(...): Can only update a mounted or mounting component. This usually means you called setState() on an unmounted component. This is a no-op. Please check the code for the LightboxOverlay component."

Because the onClose() will unmount the component, and the setState() will be illegal. 

By switching two of them, you will get no warning.