Allow custom control over the LightboxOverlay rendering.
Right now the lightbox has two methods of rendering: `Navigator` and
`Modal`.

Sometimes neither of these options work well. For example, if a client
is using `NavigationExperimental` and doesn't want to use a modal, they
are stuck.

This change allows clients to have full control over rending the
Overlay.

`showOverlay` is a callback that accepts the `LightboxOverlay` component
and associated props and clients can render the component however they
want.

`hideOverlay` is a callback so clients can specify how to hide the
overlay in a custom fashion.