renderContent can use activeProps variable
Previously, there was no way how to simply pass arguments from <Lightbox> to
the renderContent(). The only possibility was to create a new component and
utilize activeProps what works fine for simpler cases. But this component has
to be either native or it have to forward setNativeProps. Such forwarding
may be a little more complex as shown at
https://github.com/facebook/react-native/issues/1040
where change of application logic was the solution. Such solution does not
make sense with <Lightbox>.