Add `useNativeDriver` as a configuration option
Using this library the latest version of React Native (0.63.4) raises
warnings about explicitly setting `useNativeDriver` for the `Animated`
functions.

I considered hardcoding the value as false for instances of
`Animated.event` and `Animated.spring` to maintain the existing
behaviour but felt like exposing it as an option adds very little extra
complexity to the library while ensuring people can opt in to the native
animations if they want.